test: fold German/Spanish/Japanese locale tests into i18n_integration_test.rs
Now that DashboardSettings is injected instead of read from a global CONFIG
singleton, tests no longer need to live in separate binaries to get one
locale per process. Consolidate fr/de/es/ja into parallel #[tokio::test]
functions in a single file.
